Fix foreign keys using unsignedBigInteger

+++ database/migrations/2015_05_15_000080_create_user_usergroups_table.php
@@ -29,10 +29,10 @@ class CreateUserUsergroupsTable extends Migration
     public function up()
     {
         Schema::create('user_usergroup', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('usergroup_id')->unsigned();
+            $table->unsignedBigInteger('usergroup_id')->unsigned();
             $table->foreign('usergroup_id')->references('id')->on('usergroups')->onDelete('cascade');
             $table->unique(array('user_id', 'usergroup_id'));
         });
